* commit 'b673d7b0378f3ce80bcd1117af072264e7a35c15':
  increase numeric data types
  change line of code type to long

+++ api/src/main/java/com/ripstech/apiconnector2/Api.java
@@ -195,7 +195,7 @@ public class Api {
 		return service;
 	}
 
-	public ApplicationPath application(int applicationId) {
+	public ApplicationPath application(long applicationId) {
 		ApplicationPath path = new ApplicationPath(baseUri, applicationId);
 		path.setHttpClientConfig(httpClientConfig);
 		return path;
